Move creatable broadcasted notification types to /models
