+++ src/models/notification/index.ts
@@ -7,8 +7,11 @@
  **************************************************************************/
 
 export * from './broadcasted-notification';
+export * from './creatable-broadcasted-notification';
 export * from './notification';
 export * from './raw-broadcasted-notification';
+export * from './raw-creatable-broadcasted-notification';
 export * from './raw-notification';
 export * from './raw-targeted-notification';
 export * from './targeted-notification';
+export * from './to-raw-creatable-broadcasted-notification';
